requirements.txt
django
django-heroku
pillow
gunicorn
djangorestframework
django-cors-headers
dj_database_url
dropbox
django-rest-auth
